@@ -3,22 +3,24 @@
   "description": "Generate critical CSS for POSTed HTML and CSS",
   "version": "0.1.0",
   "private": true,
+  "type" : "module",
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
     "live": "nodemon index.js"
   },
   "engines": {
-    "node": ">= 12.0.0"
+    "node": ">=14.16"
   },
   "author": "TinyBit, LLC",
   "license": "MIT",
   "dependencies": {
-    "critical": "^4.0.1",
-    "express": "^4.18.1",
+    "critical": "^5.0.4",
+    "express": "^4.18.2",
+    "puppeteer": "^19.4.1",
     "tmp": "^0.2.1"
   },
   "devDependencies": {
-    "nodemon": "^2.0.19"
+    "nodemon": "^2.0.20"
   }
 }
