@@ -1,9 +1,11 @@
-const express = require('express');
-const critical = require('critical');
-const puppeteer = require('puppeteer');
-const fs = require('fs');
-const tmp = require('tmp');
+import express from 'express';
+import { generate } from 'critical';
+import puppeteer from 'puppeteer';
+import fs from 'fs';
+import tmp from 'tmp';
+
 const app = express();
+
 app.use(express.json({limit: '10mb'}));
 
 app.get('/', (req, res) => {
@@ -37,7 +39,7 @@ app.post('/', async (req, res) => {
   ];
   try {
     await fs.promises.appendFile(cssFile, req.body.css);
-    const { css, html, uncritical } = await critical.generate({
+    const { css, html, uncritical } = await generate({
       concurrency: 1, // https://github.com/addyosmani/critical/issues/364#issuecomment-493865206
       css: cssFile,
       html: req.body.html,
